Added basic agent class
